requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
qdrant-client
python-multipart
pydantic
pytest
pytest-cov
httpx